move generateVariantRecordsWithColorVariations in jobHelper
