+++ cartridges/int_algolia/cartridge/scripts/algolia/steps/algoliaProductDeltaIndex.js
@@ -274,7 +274,7 @@ exports.process = function(cpObj, parameters, stepExecution) {
             }
             if (product.master) {
                 var processedProducts = 0;
-                var recordsPerLocale = modelHelper.generateVariantRecordsWithColorVariations({
+                var recordsPerLocale = jobHelper.generateVariantRecordsWithColorVariations({
                     masterProduct: product,
                     locales: siteLocales,
                     attributeList: attributesToSend,
